UI added and other fixes and improvements

+++ src/game/RogueliteGame.java
@@ -4,7 +4,7 @@ import framework.Game;
 import framework.Scene;
 
 public class RogueliteGame extends Game {
-    public static final double gameScale = 2;
+    public static final int gameScale = 2;
     public static final int tileSize = 16;
 
     public RogueliteGame() {
